+++ driven/storage/adapter_v3.go
@@ -135,6 +135,12 @@ func (sa *Adapter) FindGroupsV3(context TransactionContext, clientID string, fil
 			{Key: "title", Value: 1},
 		})
 	}
+
+	if filter.DaysInactive != nil {
+		pastTime := time.Now().Add(-time.Duration(*filter.DaysInactive) * 24 * time.Hour)
+		groupFilter = append(groupFilter, primitive.E{Key: "date_updated", Value: bson.M{"$lt": pastTime}})
+	}
+
 	if filter.Limit != nil {
 		findOptions.SetLimit(*filter.Limit)
 	}
@@ -280,7 +286,7 @@ func (sa *Adapter) CreatePendingMembership(clientID string, user *model.User, gr
 			case "pending":
 				return errors.New("the user is pending for the group")
 			case "rejected":
-				return errors.New("the user is rejected for the group")
+				return errors.New("the user is denied for the group")
 			default:
 				return errors.New("error creating a pending user")
 			}
@@ -474,19 +480,40 @@ func (sa *Adapter) CreateMembership(clientID string, current *model.User, group 
 // CreateMemberships Created multiple members to a group
 func (sa *Adapter) CreateMemberships(context TransactionContext, clientID string, current *model.User, group *model.Group, memberships []model.GroupMembership) error {
 	now := time.Now()
+	upsert := true
 
-	var objects []interface{}
+	var objects []mongo.WriteModel
 	for index := range memberships {
-		memberships[index].ID = uuid.NewString()
-		memberships[index].ClientID = clientID
-		memberships[index].DateCreated = now
 		if memberships[index].UserID != "" && memberships[index].ExternalID != "" && memberships[index].Email != "" && memberships[index].Status != "" {
-			objects = append(objects, memberships[index])
+			objects = append(objects, &mongo.UpdateOneModel{
+				Filter: bson.M{
+					"group_id": group.ID,
+					"user_id":  memberships[index].UserID,
+				},
+				Update: bson.M{
+					"$setOnInsert": bson.M{
+						"_id":          uuid.NewString(),
+						"client_id":    clientID,
+						"group_id":     group.ID,
+						"user_id":      memberships[index].UserID,
+						"external_id":  memberships[index].ExternalID,
+						"net_id":       memberships[index].NetID,
+						"name":         memberships[index].Name,
+						"email":        memberships[index].Email,
+						"date_created": now,
+					},
+					"$set": bson.M{
+						"date_updated": now,
+						"status":       memberships[index].Status,
+					},
+				},
+				Upsert: &upsert,
+			})
 		}
 	}
 
 	if len(objects) > 0 {
-		_, err := sa.db.groupMemberships.InsertManyWithContext(context, objects, nil)
+		_, err := sa.db.groupMemberships.BulkWriteWithContext(context, objects, nil)
 		return err
 	}