+++ CHANGELOG.md
@@ -5,6 +5,35 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [1.65.0] - 2025-06-10
+### Added
+- Add new group type - "administrative" bool [#576](https://github.com/rokwire/groups-building-block/issues/576)
+
+## [1.64.3] - 2025-05-27
+### Fixed
+- Group Membership - Rejected vs Denied and display discrepancy [#574](https://github.com/rokwire/groups-building-block/issues/574)
+- Fix documentation issue [#572](https://github.com/rokwire/groups-building-block/issues/572)
+
+
+## [1.64.2] - 2025-05-21
+### Fixed
+- Additional fix on validation [#569](https://github.com/rokwire/groups-building-block/issues/569)
+
+## [1.64.1] - 2025-05-21
+### Fixed
+- Rename param name to post_update in UpdateGroupDateUpdated api [#569](https://github.com/rokwire/groups-building-block/issues/569)
+
+## [1.64.0] - 2025-05-15
+### Added
+- Add a description field to the /gr/api/analytics/groups analytics API [#567](https://github.com/rokwire/groups-building-block/issues/567)
+- Inactive Groups [#552](https://github.com/rokwire/groups-building-block/issues/552)
+
+## [1.63.0] - 2025-05-09
+### Changed
+- Improve PUT group/{group-id}/members/multi-create [#562](https://github.com/rokwire/groups-building-block/issues/562)
+### Added
+- Improve POST api/groups [#561](https://github.com/rokwire/groups-building-block/issues/561)
+
 ## [1.62.0] - 2025-04-17
 ### Changed
 - Support Google Trust Services as CA [#558](https://github.com/rokwire/groups-building-block/issues/558)